One (hopefully) last tweak.
We now start the mixing function by xoring two of the state elements
with gibberish numbers.  The main purpose of this is to protect against
data that might be generated with a similar algorithm as the reverse of
the mixing function.  In practice, that's extremely unlikely to happen
anyway.  But this has very little performance impact and provides a bit
more peace of mind, as well as better formal justification of the hash
design.

As a bonus, this also makes the mixing function no longer
zero-sensitive.  That wasn't necessary for hash quality anyway given
other aspects of the hash, but it does take the burden off of those
other aspects to compensate for that zero-sensitivity.