I lied, one more tweak: new rotation constants.
This wasn't necessary at all for the hash itself, but rather is
motivated by wanting to be as transparent as possible.  The code I used
to generate the old rotation constants is on a computer that I don't
currently have access to (I haven't shipped it to where I'm living yet).
I thought I had put the code online, but apparently not.

So I wrote a new program (also included in this commit) to optimize
constants, and switched to constants generated by that program.  This
way all code related to TentHash is out in the open and can be examined.

As a bonus, the new constants actually have even better diffusion than
the old ones.  Although that's not meaningful, since the old constants
already had more than enough diffusion for the TentHash's output size.